% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/longurl.r
\name{known_services}
\alias{known_services}
\title{Retrieve all the URL shortener services known to the longurl API}
\usage{
known_services()
}
\description{
Retrieve all the URL shortener services known to the longurl API
}
\examples{
short_svcs <- known_services()
head(short_svcs)
}

