% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/longurl-package.r
\docType{package}
\name{longurl}
\alias{longurl}
\alias{longurl-package}
\title{A package to work with the longurl API}
\description{
A package to work with the longurl API
}
\author{
Bob Rudis (@hrbrmstr)
}

