% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/longurl-package.r
\docType{package}
\name{longurl}
\alias{longurl}
\alias{longurl-package}
\title{A package to work with the LongURL API}
\description{
Interface to the 'LongURL' API to identify known URL shortener services
and expand vectors of short URLs with optional error checking and
URL validation.
}
\author{
Bob Rudis (@hrbrmstr)
}

