% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longurl-package.r
\docType{package}
\name{longurl}
\alias{longurl}
\alias{longurl-package}
\title{A package to work with the URL-Expander API}
\description{
Interface to the 'URL-Expander' API to expand vectors of short URLs
with optional error checking and URL validation.
}
\author{
Bob Rudis (bob@rud.is)
}

