% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longurl-package.r
\docType{package}
\name{longurl}
\alias{longurl}
\alias{longurl-package}
\title{Tools expand vectors of short URLs into long URLs}
\description{
Tools expand vectors of short URLs into long URLs
}
\author{
Bob Rudis (bob@rud.is)
}
